+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## master (unreleased)
 
+- Add support for resolvers. ([@palkan][])
+
+Now it's possible to `include ActionPolicy::GraphQL::Behaviour` into resolver class to use
+Action Policy helpers there.
+
 ## 0.1.0 (2019-05-20)
 
 - Initial version. ([@palkan][])